fix: remove all non-breaking spaces
Found a lot of non-breaking spaces in places where they were not needed.
Replace them with spaces and fix the resulting whitespace errors.

Out of curiosity I was scanning the docs for any non-ascii characters to
see if there is anything that could be breaking formatting that standard
mechanisms wouldn't catch. Found all of this. Assume it was from the
initial import from the processor wiki.

Signed-off-by: Randolph Sapp <rs@ti.com>