feat(cicd): automated builds and deployment to pages
Use a matrix to parallelize building the docs for listed DEVFAMILY and
OS combinations. Aggregate the results to build an expected pages
artifact. Deploy the monolithic artifact to GitHub pages.

If any platform fails for any reason, the arguments and logs for that
build will be logged under a separate section of this workflow, and the
results should not be deployed.

Signed-off-by: Randolph Sapp <rs@ti.com>