feat(android): Document audio capture bug
We received an e2e error for an audio capture bug.

Document the steps that are needed to solve it by linking to
the patches that need to be applied.

Note that for the tinyhal patch, we cannot link to git.ti.com
because the project does not exist there.

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>