feat: linux: Wave5: Add Section for Stream Quality Improvements
There have been a handful of customer questions on video quality lately
that has led to repeating similar answers across different channels. As
a result, a section should be made that can be constantly added to with
any recommendations on improving video quality. Initially, this section
is being made with the recommendations being offered most often.

Signed-off-by: Brandon Brnich <b-brnich@ti.com>