chore: move to include only logic
For a while this project mixed including and excluding document logic,
where the include list was an exhaustive description of everything
needed by default.

Because this list is already exhaustive we can drop the conditional
exclude logic, exclude everything by default, and use the include list
to control documents we parse.

Signed-off-by: Randolph Sapp <rs@ti.com>