chore(j7-family): finally split the j7 family
This used to be convenient, but ultimately there's no guarantee a device
stays part of a family or not. On top of that, there should be no reason
for Jacinto to be special here. Finally split the family so we can
remove some exceptions to general rules other platforms follow.

Signed-off-by: Randolph Sapp <rs@ti.com>