ci(vale): add a pr check workflow for vale
Add a pull request driven workflow for vale. This is based on the
upstream action documentation [1], but limit it to changed files only to
keep feedback snappy.

This currently uses a fork I made of the upstream action that's capable
of using the locally installed version of vale. This allows us to use
our container as is, which is faster than reinstalling tools and makes
issues more easily reproducible. This is only required until the
upstream PR is accepted [2].

This does not enable this action as a blocking check just yet. There is
still an upstream bug that makes all feedback throw a failure, which is
a little too aggressive right now. Still need to tune things a bit.

[1] https://github.com/errata-ai/vale-action
[2] https://github.com/errata-ai/vale-action/pull/149

Signed-off-by: Randolph Sapp <rs@ti.com>