feat(dedupe): add a deduplication script
Add a script capable of deduplicating assets used between multiple build
targets.

When attempting to play with sphinx's asset paths I noticed there was a
lot of logic around preserving relative links. It didn't seem like they
want to allow usage of an asset path outside of the build directory.

The WebSupport class seemed to allow some of that flexibility we were
looking for, but unfortunately it still assumes the conf.py is in the
source directory and has issues working with other Sphinx extensions
like ifconfig.

Given what we are doing seems to be a niche usecase, we can address this
with a little post processing using lxml filters.

Signed-off-by: Randolph Sapp <rs@ti.com>