fix(release-notes): use release-specific-build-information
Use the common release-specific-build-information tag instead of the
component specific tags. This data was always somewhat tabular but was
split over several sections with different formatting depending on the
platform. We are now looking to merge this under a single table like
what was done with AM62LX.

This initial merge made some links inaccessible for some platforms. This
attempts to go back and enforce a common tag name to make sure we are
always linking the user to at least a close section while the transition
occurs.

Signed-off-by: Randolph Sapp <rs@ti.com>