fix(linux): Update ICSSM FAQ on how to address low performance
In case of AM335x and AM437, the TCP performance numbers tend
to be low, where the bottle neck is the CPU frequency and the
throughput attained in the expected one. Explain the same in the
FAQ of ICSSM doc and also provide the command which can be used
if the user wants to attain max possible throughput.

Signed-off-by: Meghana Malladi <m-malladi@ti.com>